requirements.txt
allure-pytest==2.8.12
allure-python-commons==2.8.12
future==0.18.2
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.51.0
grpcio
paho-mqtt==1.5.0
paramiko==2.7.1
protobuf==3.11.3
psutil
pyserial==3.4
pytest==5.4.3
pytest-forked==1.1.3
pytest-logger==0.5.1
pytest-repeat==0.8.0
pytest-rerunfailures==9.0
pytest-timeout==1.3.4
pytest-xdist==1.31.0
python-dateutil==2.8.1
python-vxi11==0.9
robotframework==3.1.2
robotframework-appiumlibrary==1.5.0.4
robotframework-mqttlibrary==0.7.0
robotframework-seleniumlibrary==4.3.0
robotframework-sshlibrary==3.4.0
singleton-decorator==1.0.0
SpeechRecognition==3.8.1
urllib3==1.25.8
xlrd==1.2.0
zipp==3.1.0
zope.interface==5.0.2
artifactory
pywin32
execnet
pebble
vbf-parser
pythonnet